timer0 cont mode

+++ adc_dac.h
@@ -1,5 +1,5 @@
 #include <stdint.h>
-void dac_set_voltage(int8_t voltage);
+void dac_set_voltage(uint16_t voltage);
 void adc_init(void);
 uint16_t adc_read_current(void);
 uint16_t adc_read_voltage(void);
